use doctest to launch benchmarks
